+++ .vscode/launch.json
@@ -1,0 +1,12 @@
+{
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "Run calculator script",
+            "type": "python",
+            "request": "launch",
+            "program": "Calculator/calculator.py",
+            "console": "integratedTerminal",
+        }
+    ]
+}
