+++ Calculator/calculator.py
@@ -1,22 +1,18 @@
-from numbers import Real
+from sum import sum
+from div import div
+import times
 
-def sum(num1, num2):
-    # Check if both inputs are numbers
-    if isinstance(num1, Real) and isinstance(num2, Real):
-        return num1 + num2
-    else:
-        print("This method only allows numeric values.")
+print("This will be our new calculator")
+print("Now we have submodules")
 
-def times(num1, num2):
-    # Check if both inputs are numbers
-    if isinstance(num1, Real) and isinstance(num2, Real):
-        return num1 * num2
-    else:
-        print("This method only allows numeric values.")
-                
-def subs(num1, num2):
-    # Check if both inputs are numbers
-    if isinstance(num1, Real) and isinstance(num2, Real):
-        return num1 - num2
-    else:
-        print("This method only allows numeric values.")
+print("-----------------------------------")
+print("Our first operation will be a sum")
+print("How much is 1 + 2 = " + sum(1,2))
+
+print("-----------------------------------")
+print("Now let's do a times operation")
+print("How much is 3 * 1 = " + str(times.times(1,3)))
+
+print("-----------------------------------")
+print("Now let's do a div")
+print("How much is 3 * 1 = " + div(1,0))
