+++ Calculator/subs.py
@@ -1,0 +1,8 @@
+from numbers import Real
+
+def subs(num1, num2):
+    # Check if both inputs are numbers
+    if isinstance(num1, Real) and isinstance(num2, Real):
+        return num1 - num2
+    else:
+        print("This method only allows numeric values.")
